Refactor namespace and examples.
