+++ examples/convert_eeg_to_bids.py
@@ -34,8 +34,7 @@ import shutil as sh
 import mne
 from mne.datasets import eegbci
 
-from mne_bids import write_raw_bids, make_bids_basename
-from mne_bids.path import print_dir_tree
+from mne_bids import write_raw_bids, make_bids_basename, print_dir_tree
 
 ###############################################################################
 # Download the data
