+++ examples/convert_empty_room.py
@@ -27,8 +27,8 @@ from datetime import datetime, timezone
 import mne
 from mne.datasets import sample
 
-from mne_bids import write_raw_bids, read_raw_bids, make_bids_basename
-from mne_bids.path import print_dir_tree
+from mne_bids import (write_raw_bids, read_raw_bids,
+                      make_bids_basename, print_dir_tree)
 
 ###############################################################################
 # And define the paths and event_id dictionary.
