+++ examples/convert_mri_and_trans.py
@@ -47,8 +47,7 @@ from mne.datasets import sample
 from mne.source_space import head_to_mri
 
 from mne_bids import (write_raw_bids, make_bids_basename, write_anat,
-                      get_head_mri_trans)
-from mne_bids.path import print_dir_tree
+                      get_head_mri_trans, print_dir_tree)
 
 ###############################################################################
 # We will be using the `MNE sample data <mne_sample_data_>`_ and write a basic
